+++ internal/framework/validation/objectvalidation/exactly_one_of_if_object.go
@@ -1,4 +1,4 @@
-package validation
+package objectvalidation
 
 import (
 	"context"